Artificial Intelligence/Advanced/LLM (LANGCHAIN GEN AI)/LLM3-Gemini Pro LLM Application/requirements.txt
streamlit
google-generativeai
python-dotenv